renamed the files back to what they were
